Fully converted Stocks Page
Stocks page has been fully converted to tailwind, and has a similar view
to the original CSS. The only real change in the html other than adding
TW classes is the span in the search results was removed (since it
didn't seem to be doing anything anyway). The hx-indicator was also
removed from the Add button, if/when I decide to reincorporate that I'll
need to spend more time understanding how to handle the indicator w/
tailwind.